requirements.txt
pandas 
openpyxl
duckdb
python-dotenv
rich
openai